requirements.txt
# Roboflow API
roboflow>=1.1.0

# Data handling
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0

# COCO format handling
pycocotools>=2.0.6

# Configuration
PyYAML>=6.0

# Deep Learning - SAM3 Training
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0
timm>=0.9.0

# Training utilities
tqdm>=4.65.0
tensorboard>=2.14.0
scikit-learn>=1.3.0

# Model evaluation
torchmetrics>=1.2.0

# Experiment tracking
mlflow>=2.9.0
